Move HBase endpoint validators to utility
